Fix bug where only first file was extracted when using -x to split epub

+++ tests/test-epub-merge.sh
@@ -1,51 +1,73 @@
 #!/bin/bash
 
-# WARNING: Local environment only - do not run elsewhere
+# WARNING: Local use only - do not run elsewhere
 
 set -euo pipefail
 trap 'echo "Error at line $LINENO" >&2' ERR INT TERM
 
+if [[ $# -eq 1 && ( "$1" = "clean" || "$1" = "clear" ) ]]; then
+	rm -rf "$HOME/Test/epub-test/.merged/" \
+		"$HOME/Test/epub-test/.merged-splitted/" \
+		"$HOME/Test/epub-test/.splitted/"
+fi
+
 tcd() {
+	[[ -d ~/Test/epub-test/"$1"/ ]] && exit
 	rm -rf ~/Test/epub-test/"$1"/
 	mkdir -p ~/Test/epub-test/"$1"/
 	cd ~/Test/epub-test/"$1"/
 }
 
-echo ++ test unit: setup
+(
+	[[ -d ~/Test/epub-test ]] && exit
+	echo ++ Test unit: setup
 
-mkdir -p ~/Test/epub-test
-rsync -a --delete /Volumes/Norway/Backup/Test/epub-test/ ~/Test/epub-test
+	mkdir -p ~/Test/epub-test
+	rsync -a --delete --exclude .merged --exclude .merged-splitted \
+		--exclude .splitted /Volumes/Norway/Backup/Test/epub-test/ \
+		~/Test/epub-test
+)
 
-echo ++ test unit: merge
+(
+	tcd .merged
+	echo ++ Test unit: merge
 
-tcd .merged
+	ls ../merged | sed -e 's/.epub//' | while read -r line; do
+		epub-merge -q ../original/"$line"*.epub
+	done
 
-ls ../merged | sed -e 's/.epub//' | while read -r line; do
-	epub-merge -q ../original/"$line"*.epub
-done
+	cd merged
+	for i in *.epub; do
+		epub-diff.sh "$i" ../.merged/"$i"
+	done
+)
 
-for i in *.epub; do
-	epub-diff.sh "$i" ../merged/"$i"
-done
+(
+	tcd .splitted
+	echo ++ Test unit: split
 
-echo ++ test unit: split
+	for i in ../.merged/*; do
+		epub-merge -q -x "$i"
+	done
 
-tcd .splitted
+	cd ../splitted
+	for i in *.epub; do
+		epub-diff.sh "$i" ../.splitted/"$i"
+	done
+)
 
-for i in ../merged/*; do
-	epub-merge -q -x "$i"
-done
+(
+	tcd .merged-splitted
+	echo ++ Test unit: merge-splitted
 
-for i in *.epub; do
-	epub-diff.sh "$i" ../splitted/"$i"
-done
+	ls ../merged | sed -e 's/.epub//' | while read -r line; do
+		epub-merge -q ../splitted/"$line"*.epub
+	done
 
-echo ++ test unit: merge-splitted
+	cd merged
+	for i in *.epub; do
+		epub-diff.sh "$i" ../.merged-splitted/"$i"
+	done
+)
 
-tcd .merged-splitted
-
-ls ../merged | sed -e 's/.epub//' | while read -r line; do
-	epub-merge -q ../splitted/"$line"*.epub
-done
-
-echo All done
+echo "++ All done"
